Added groundwork for meta-keys, added css back in
There looks to be a section for meta keys in both the source and in the demo
area, but we added the functionality in the demo area. haven't tested any yet,
however.

also added the css back in in index, renamed, began to modify it to look like a
keyboard instead of a numpad. the template is still all the same; need to
review what classes are availiable and what we need.